requirements.txt
# TensorLake Application Requirements
# These dependencies are installed in the custom image at deployment time

tensorlake>=0.1.0
google-genai
requests>=2.31.0
psycopg2-binary>=2.9.0
